requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-django==4.5.2
gunicorn==20.1.0
dj-database-url==2.0.0
whitenoise==6.4.0
Brotli==1.0.9
pytest-cov==4.0.0
flake8==6.0.0
black==23.3.0
isort==5.12.0
